Make token validation modular
