+++ internal/output/identity/output.go
@@ -7,8 +7,8 @@ import (
 	"github.com/robbiemcmichael/auth-mux/internal/types"
 )
 
-type Output struct{}
+type Config struct{}
 
-func (o *Output) Handler(w http.ResponseWriter, validation types.Validation) error {
+func (c *Config) Handler(w http.ResponseWriter, validation types.Validation) error {
 	return json.NewEncoder(w).Encode(validation)
 }
